feat: add types to repository classes
